chore: define test models once to ease reading & authoring tests
